Veda_Distributed_Reasoning_Agent/requirements.txt
langchain
langchain_groq
google-generativeai
python-dotenv
rich
